requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.60.0
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.0
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
lap==0.5.12
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
polars==1.33.1
protobuf==6.32.1
psutil==7.1.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.202
ultralytics-thop==2.0.17
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
